Moved layouts and charts to folder, changed 15 to 14 days
