+++ src/app/shared/charts/line-chart/line-chart.component.html
@@ -1,0 +1,17 @@
+<ngx-charts-area-chart 
+    [view]="view" 
+    [scheme]="colorScheme" 
+    [results]="data" 
+    [gradient]="gradient" 
+    [xAxis]="showXAxis"
+    [yAxis]="showYAxis" 
+    [legend]="showLegend" 
+    [showXAxisLabel]="showXAxisLabel" 
+    [showYAxisLabel]="showYAxisLabel"
+    [xAxisLabel]="xAxisLabel" 
+    [yAxisLabel]="yAxisLabel" 
+    (select)="onSelect($event)"
+    [yAxisTickFormatting]="yAxisTickFormatting" 
+    [trimYAxisTicks]="trimYAxisTicks" 
+    [timeline]="timeline">
+</ngx-charts-area-chart>
