+++ src/app/shared/charts/line-chart/line-chart.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { LineChartComponent } from './line-chart.component';
 import { NgxChartsModule } from '@swimlane/ngx-charts';
-import { SharedModule } from '../shared.module';
+import { SharedModule } from '../../shared.module';
 import { CommonModule } from '@angular/common';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
