+++ src/app/shared/layouts/layout.directive.spec.ts
