+++ src/app/shared/layouts/layout.directive.ts
