+++ src/app/shared/layouts/two-by-two-layout/two-by-two-layout.component.html
@@ -15,6 +15,4 @@
             <ng-template appChart></ng-template>
         </div>
     </div>
-
-
 </div>
