+++ src/app/shared/shared.module.ts
@@ -2,12 +2,12 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 import { WidgetComponent } from './widget/widget.component';
-import { TwoByTwoLayoutComponent } from './two-by-two-layout/two-by-two-layout.component';
-import { LayoutComponent } from './layout/layout.component';
-import { ChartDirective } from './chart.directive';
-import { ChartComponent } from './chart/chart.component';
-import { LayoutDirective } from './layout.directive';
-import { LineChartComponent } from './line-chart/line-chart.component';
+import { TwoByTwoLayoutComponent } from './layouts/two-by-two-layout/two-by-two-layout.component';
+import { LayoutComponent } from './layouts/layout/layout.component';
+import { ChartDirective } from './charts/chart.directive';
+import { ChartComponent } from './charts/chart/chart.component';
+import { LayoutDirective } from './layouts/layout.directive';
+import { LineChartComponent } from './charts/line-chart/line-chart.component';
 import { NgxChartsModule } from '@swimlane/ngx-charts';
 
 @NgModule({
